feat: add loader demo to usage guide
Added a new demo component to demonstrate the usage of loaders in Vue3-Pixi. The loader component uses the `Loader` class to load a list of images and show a loading screen until all images have finished loading. The demo component uses the `Application` component from Vue3-Pixi to create a PixiJS application and render the loaded images as sprites.