docs: add CODE_OF_CONDUCT.md
Addition of a Code of Conduct document outlining the standards of behavior expected of contributors to the project in line with the Contributor Covenant. These standards apply both within project spaces and in public spaces when representing the project or its community. Project maintainers are responsible for enforcing these standards and taking appropriate corrective action in response to any unacceptable behavior.