chore: refactor pixi.js patchProp function
Refactor the `patchProp` function in `patchProp.ts` to remove the use of `camelize` and instead access the `patchProp` function of specific renderers in `utils.ts`, allowing for better handling of prop names that need to be in a specific case/format. Removed the `patchProps` array as it is no longer needed with this change.