feat(components): add External component
Added a new component called External, which allows rendering Vue components outside of the PixiJS application. This component receives props for the HTML tag to use and the root element where the component will be mounted. The implementation includes creating a new Vue application and inheriting global properties from the parent application.