docs(guide): add initialization steps for Vue plugin
This commit adds initialization steps for the Vue plugin in the quick-start and introduction sections of the documentation. The steps include importing `compilerOptions` and `transformAssetUrls` from `vue3-pixi` and adding them to the `template` property of the `vue` plugin configuration in `vite.config.js`. These steps support custom elements, prevent parsing exceptions, and enable parsing of the texture attribute, just like the `src` attribute on an image.